requirements.txt
asgiref==3.7.2
Django==4.2.7
django-debug-toolbar==4.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.20
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2023.3
